64 bit corrections to parameters to OnUpdateInt Bug #20994 (Dave)
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@119719 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/pfpro/php_pfpro.h
@@ -54,10 +54,10 @@ PHP_FUNCTION(pfpro_process);            /* Transaction processing     */
 ZEND_BEGIN_MODULE_GLOBALS(pfpro)
 	int initialized;
 	char *defaulthost;
-	int defaultport;
-	int defaulttimeout;
+	long defaultport;
+	long defaulttimeout;
 	char *proxyaddress;
-	int proxyport;
+	long proxyport;
 	char *proxylogon;
 	char *proxypassword;
 ZEND_END_MODULE_GLOBALS(pfpro)
